+++ client/components/blocks/HeaderBlock.jsx
@@ -1,20 +1,26 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
+import { Condition } from '../../modules/components'
+import clientConfig from '../../config/config';
+
+/**
+ * Header display containing navigation links
+ */
 class HeaderBlock extends React.Component {
 
   constructor(props) {
     super(props);
 
+    // Set mobile nav menu to closed initially
     this.state = {
-      isMobileNavigationToggled: true,
-      mobileNavigationClass: 'mobile-navigation'
+      isMobileNavOpen: false
     }
   }
 
   /**
    * Determines if a particular header link is disabled.
-   * @param link
+   * @param name
    */
   isDisabled = (name) =>
     this.props.disabled && this.props.disabled[name.toLowerCase()];
@@ -23,27 +29,24 @@ class HeaderBlock extends React.Component {
    * Sets the mobile navigation display
    * @param event
    */
-  handleMobileNaviconClick = (event) => {
+  toggleMobileNav = (event) => {
     event.preventDefault();
-    if (this.state.isMobileNavigationToggled) {
-      this.setState({
-        isMobileNavigationToggled: false,
-        mobileNavigationClass: 'mobile-navigation active'
-      });
-    } else {
-      this.setState({
-        isMobileNavigationToggled: true,
-        mobileNavigationClass: 'mobile-navigation'
-      });
-    }
+    this.setState({ isMobileNavOpen: !this.state.isMobileNavOpen });
   };
 
+  /**
+   * Helper to render a icon/link combination
+   * @param {object} options
+   */
   renderLink = (options) => {
     const isDisabled = this.isDisabled(options.name);
     return (isDisabled) ? (
       <li className='disabled'>
         <a>
           <i className={options.iconClass} />
+          <Condition condition={this.state.isMobileNavOpen}>
+            <span className="navigation-tooltip disabled">{options.name}</span>
+          </Condition>
         </a>
       </li>
     ) : (
@@ -56,12 +59,20 @@ class HeaderBlock extends React.Component {
     );
   };
 
+  /**
+   * Helper to render a icon/link combination. Different from renderLink
+   * in that these items have an onClick action when enabled.
+   * @param {object} options
+   */
   renderActionLink = (options) => {
     const isDisabled = this.isDisabled(options.name);
     return (isDisabled || !options.action) ? (
       <li className='disabled'>
         <a>
           <i className={options.iconClass} />
+          <Condition condition={this.state.isMobileNavOpen}>
+            <span className="navigation-tooltip disabled">{options.name}</span>
+          </Condition>
         </a>
       </li>
     ) : (
@@ -81,21 +92,21 @@ class HeaderBlock extends React.Component {
       <div className="header-block unselectable">
         <Link to="/">
           <div className="left-nav">
-            <img src="/img/logo.svg" alt="logo" className="logo" />
-            <h1>{document.title}</h1>
+            <img src={clientConfig.logoUrl} alt="logo" className="logo" />
+            <h1>{clientConfig.headerTitle}</h1>
           </div>
         </Link>
 
         <div className="right-nav">
 
           <div className="mobile-navicon">
-            <a href="#" onClick={this.state.handleMobileNaviconClick}>
+            <a href="#" onClick={this.toggleMobileNav}>
               <i className="fa fa-bars" />
             </a>
           </div>
 
           { /** Mobile Navigation **/ }
-          <ul className={this.state.mobileNavigationClass}>
+          <ul className={`mobile-navigation ${(this.state.isMobileNavOpen) ? ' active' : ''}`}>
             {this.renderLink({to: '/', iconClass: 'fa fa-plus', name: 'New'})}
             {this.renderActionLink({ action: this.props.saveButton, name: 'Save', iconClass: 'fa fa-floppy-o' })}
             {this.renderLink({to: `/edit`, iconClass: 'fa fa-pencil', name: 'Edit', state: { editLink: true, currentPaste: this.props.currentPaste }})}