+++ client/components/views/HelpView.jsx
@@ -25,13 +25,18 @@ class HelpView extends React.Component {
 
   render() {
     return (
-      <div className={`help-view flex-container ${this.context.styleStore.theme}`}>
+      <div className={
+        `help-view flex-container ${this.context.styleStore.theme.className}`
+      }>
         <HeaderBlock
           disabled={{ raw: true, edit: true, save: true }}
         />
         <div className="view-container">
           <div>
-            Please submit bug reports or feature requests to: <a href="https://github.com/evermoredev/everpaste/issues">EverPaste Repo</a>
+            Please submit bug reports or feature requests to:
+            <a href="https://github.com/evermoredev/everpaste/issues">
+              EverPaste Repo
+            </a>
           </div>
           <div
             className="code-document"