+++ client/components/views/PublicListView.jsx
@@ -12,7 +12,7 @@ class PublicListView extends React.Component {
     this.state = {
       docList: [],
       loading: true
-    }
+    };
   }
 
   // Once the component mounts, make the api request for the list of docs
@@ -39,7 +39,9 @@ class PublicListView extends React.Component {
         <div className="public-paste">
           <Condition value={d.title} className="title" default="Untitled" />
           <Condition value={d.name} className="from-name" />
-          <div className="created">({moment(d.created).fromNow()})</div>
+          <div className="created">
+            ({moment(d.createdAt).from(moment.utc().format("YYYY-MM-DD H:mm:ss.SSS"))})
+          </div>
         </div>
       </Link>
     </div>
@@ -47,7 +49,7 @@ class PublicListView extends React.Component {
 
   render() {
     return (
-      <div className={`public-list-view flex-container ${this.context.styleStore.theme}`}>
+      <div className={`public-list-view flex-container ${this.context.styleStore.theme.className}`}>
         <HeaderBlock
           disabled={{ raw: true, edit: true, save: true }}
         />