+++ client/index.jsx
@@ -1,11 +1,14 @@
-import './styles/main.scss';
-import 'react-hot-loader/patch';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { AppContainer } from 'react-hot-loader';
-import { RootContainer } from './components/containers';
+import 'react-hot-loader/patch';
 
+import { RootContainer } from './components/containers';
+import './styles/main.scss';
+
+// Grab the root element
 const rootEl = document.getElementById('root');
+
 const render = Component =>
   ReactDOM.render(
     <AppContainer>
@@ -15,6 +18,8 @@ const render = Component =>
   );
 
 render(RootContainer);
+
+// For hot reloading
 if (module.hot)
   module.hot.accept('./components/containers/RootContainer.jsx',
     () => render(RootContainer));