+++ server/config/constants.js
@@ -1,0 +1,5 @@
+// Helper for setting the content-type
+export const contentType = {
+  json:  { 'content-type': 'application/json' },
+  plain: { 'content-type': 'text/plain' }
+};