+++ server/controllers/api_controller.js
@@ -1,28 +1,22 @@
-/**
- * Handles saving and retrieving documents from the store.
- *
- * TODO:
- *   - More error handling and logging
- */
-
-import winston from 'winston';
-import PostgresStore from './postgres_store';
-import RandomKeyGenerator from './key_generator';
 import fs from 'fs';
 import path from 'path';
-import config from '../config/config';
+import winston from 'winston';
+
+import { postgresTimestamp } from '../modules/_helpers';
+
+import { contentType } from '../config/constants';
 import { privacyOptions } from '../../shared/config/constants';
-import { postgresTimestamp } from './_helpers';
 import { fileValidation, pasteValidation } from '../../shared/validations/paste';
 import { mime2ext } from '../../shared/modules/mime2ext';
 
-class DocHandler {
+import db from '../models';
 
-  static contentType = {
-    json:  { 'content-type': 'application/json' },
-    plain: { 'content-type': 'text/plain' }
-  };
+/**
+ * Class for handling all the api calls for the pastebin
+ */
+class ApiController {
 
+  // Helper for splitting the docKey and language
   static splitDocKey = docKey => {
     const docParts = docKey.split('.', 2);
     return {
@@ -36,41 +30,38 @@ class DocHandler {
     return (fs.existsSync(filePath)) ? filePath : null;
   };
 
-  constructor() {
-    this.store = new PostgresStore();
-    this.keyGenerator = new RandomKeyGenerator();
-  }
+  constructor() { }
 
   fail = (res, opts) => {
     if (opts.serverMsg) winston.warn(opts.serverMsg);
-    res.writeHead(opts.resCode || 404, DocHandler.contentType.json);
+    res.writeHead(opts.resCode || 404, contentType.json);
     res.end(JSON.stringify({ message: opts.clientMsg }));
   };
 
   async handleGet(docKey, res, options = {}) {
-    const { key, lang } = DocHandler.splitDocKey(docKey);
-    const data = await this.store.getByKey(key);
+    const { key, lang } = ApiController.splitDocKey(docKey);
+    const data = await db.entries.getEntry(key);
 
     // Return the data object if there's text available or if it's a file paste
     // and the file still exists
-    if (data && (data.text || (data.filename && DocHandler.getFilePath(data.filename)))) {
+    if (data && (data.text || (data.filename && ApiController.getFilePath(data.filename)))) {
       winston.verbose('Retrieved document', { key });
       res.end(JSON.stringify(data));
     } else {
       winston.warn('Document not found', { key });
-      res.writeHead(404, DocHandler.contentType.json);
+      res.writeHead(404, contentType.json);
       res.end(JSON.stringify({ message: 'Document not found.' }));
     }
   }
 
   async handleRawGet(docKey, res) {
-    const { key, lang } = DocHandler.splitDocKey(docKey),
-          data = await this.store.getByKey(key);
+    const { key, lang } = ApiController.splitDocKey(docKey),
+          data = await db.entries.getEntry(key);
 
     // Don't return raw docs that are encrypted
     if (data && data.text && data.privacy != privacyOptions.encrypted) {
       winston.verbose('Retrieved raw document', { key });
-      res.writeHead(200, DocHandler.contentType.plain);
+      res.writeHead(200, contentType.plain);
       res.end(data.text);
     }
 
@@ -81,14 +72,17 @@ class DocHandler {
   }
 
   /**
-   * TODO: check the database to make sure the paste hasn't expired
-   * @param filename
+   * @param key
    * @param res
    */
-  handleGetFile(filename, res) {
-    const filePath = path.join(__dirname, `../uploads/${filename}`);
-    if (fs.existsSync(filePath)) {
-      return res.sendFile(filePath);
+  async handleGetFile(key, res) {
+    const data = await db.entries.getEntry(key);
+
+    if (data && data.filename) {
+      const filePath = path.join(__dirname, `../uploads/${data.filename}`);
+      if (fs.existsSync(filePath)) {
+        return res.sendFile(filePath);
+      }
     }
 
     this.fail(res, {
@@ -98,12 +92,18 @@ class DocHandler {
   }
 
   async handleGetList(req, res) {
-    const data = await this.store.getList();
+    const data = await db.entries.getAll();
     winston.verbose('Retrieving document list.');
-    res.writeHead(200, DocHandler.contentType.plain);
+    res.writeHead(200, contentType.plain);
     res.end(JSON.stringify(data));
   }
 
+  /**
+   * Handles post requests to /api
+   * @param req
+   * @param res
+   * @returns {Promise.<void>}
+   */
   async handlePost(req, res) {
     let data = req.body;
     data.file = req.files[0];
@@ -119,16 +119,16 @@ class DocHandler {
     }
 
     // Generate a new key
-    const key = await this.chooseKey();
+    data.key = await db.entries.createKey();
 
     // Try writing the file if it exists
     try {
       if (data.file) {
-        data.filename = key + mime2ext[data.file.mimetype];
+        data.filename = data.key + mime2ext[data.file.mimetype];
         fs.writeFileSync(`server/uploads/${data.filename}`, data.file.buffer);
       }
     } catch(e) {
-      return this.fail({
+      return this.fail(res, {
         resCode: 500,
         clientMsg: 'Problem uploading file. Please try again later.',
         serverMsg: 'File error: ' + e
@@ -141,37 +141,25 @@ class DocHandler {
       (['30 minutes', '6 hours', '1 days', '1 weeks', '1 months'].includes(data.expiration)) ?
         postgresTimestamp(data.expiration) : null;
 
-    // Insert the query
-    const queryInserted = await this.store.insert(key, data);
+    // delete the file property so it's not added in the insert
+    // delete data.file;
 
-    if (queryInserted) {
-      winston.verbose('Added document: ', { key });
-      res.writeHead(200, DocHandler.contentType.json);
-      res.end(JSON.stringify({ key }));
+    // Insert the query
+    const created = await db.entries.createSync(data);
+
+    if (created) {
+      winston.verbose('Added document: ', { key: data.key });
+      res.writeHead(200, contentType.json);
+      res.end(JSON.stringify({ key: data.key }));
     } else {
-      this.fail({
+      this.fail(res, {
         clientMsg: 'Problem adding document. Please try again later.',
-        serverMsg: 'Error adding document for key: ' + key,
+        serverMsg: 'Error adding document for key: ' + data.key,
         resCode: 500
       });
     }
   };
 
-  /**
-   * Create a random key that doesn't already exist in the database.
-   * @returns string
-   */
-  async chooseKey() {
-    let foundKey = false;
-    while (!foundKey) {
-      let key = this.keyGenerator.createKey();
-      let keyExists = await this.store.getByKey(key);
-      if (!keyExists) {
-        return key;
-      }
-    }
-  }
-
 }
 
-export default DocHandler;
+export default ApiController;