+++ shared/config/constants.js
@@ -3,3 +3,6 @@ export const privacyOptions = {
   private: 'private',
   encrypted: 'encrypted'
 };
+
+export const expirationOptions =
+  ['30 Minutes', '6 Hours', '1 Day', '1 Week', '1 Month', 'Forever'];